v1: Split handlers and middleware out to separate files
